% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerosol_world.R
\docType{data}
\name{aerosol_world}
\alias{aerosol_world}
\title{Data from of NASA Earth Observations at https://neo.gsfc.nasa.gov}
\format{
A stars object with x, y and time containing aerosol thickness. Dimensions 360x180x13
}
\usage{
aerosol_world
}
\description{
Aerosol optical thickness data from December 2019 to December 2020, taken monthly.
}
\keyword{datasets}
