% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_cannonical_correlation.R
\name{autoplot.cancor}
\alias{autoplot.cancor}
\title{Plots Canonical Correlation Analysis variables from cannonical_correlation function.}
\usage{
\method{autoplot}{cancor}(object, xlab = "Time", ...)
}
\arguments{
\item{object}{The output of the function `cannonical_correlation'.}

\item{xlab}{The xlabel to appear on CCA plot.}

\item{...}{Other arguments currently ignored.}
}
\description{
Plots Canonical Correlation Analysis (CCA) variables using 2 datasets.
}
\examples{
df1 <- SSTdatashort[1:100, ]
df2 <- SSTdatashort[401:500, ]
cca <- canonical_correlation(df1, df2)
autoplot(cca, , xlab = "Longitude")

}
