% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_emp_orth_fun.R
\name{autoplot.emporthfun}
\alias{autoplot.emporthfun}
\title{Plots empirical orthogonal function from emp_orth_fun}
\usage{
\method{autoplot}{emporthfun}(
  object,
  EOF_num = 1,
  palette = "Spectral",
  only_EOF = FALSE,
  only_TS = FALSE,
  ...
)
}
\arguments{
\item{object}{The output of the function `emp_orth_fun'.}

\item{EOF_num}{The Empirical Orthogonal Function (EOF) to plot.}

\item{palette}{The color palette. Default is \code{Spectral}.}

\item{only_EOF}{If \code{TRUE}, only the spatial EOF function would be plotted.}

\item{only_TS}{If \code{TRUE}, only the PC time series would be plotted. If both are set to \code{FALSE},
both plots would be displayed. Both cannot be set to \code{TRUE}.}

\item{...}{Other arguments currently ignored.}
}
\description{
Plots either the empirical orthogonal functions or the PC time series for EOFs.
}
\examples{
data(SSTlonlatshort)
data(SSTdatashort)
data(SSTlandmaskshort)
delete_rows <- which(SSTlandmaskshort  ==  1)
SSTdatashort   <- SSTdatashort[-delete_rows, 1:396]
emp1 <- emp_orth_fun(SSTlonlatshort[-delete_rows,  ],
                     SSTdatashort)
autoplot(emp1,
         EOF_num = 1)

autoplot(emp1,
         EOF_num = 3,
         only_EOF = TRUE)

autoplot(emp1,
         EOF_num = 2,
         only_TS = TRUE)

}
