% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_hovmoller.R
\name{autoplot.hovmoller}
\alias{autoplot.hovmoller}
\title{Hovmoller plots for latitude or longitude}
\usage{
\method{autoplot}{hovmoller}(
  object,
  ylab = "Day",
  xlab = NULL,
  title = "",
  palette = "Spectral",
  legend_title = "z",
  ...
)
}
\arguments{
\item{object}{The output of the function `hovmoller'.}

\item{ylab}{The y label.}

\item{xlab}{The x label.}

\item{title}{The graph title.}

\item{palette}{The color palette. Default is \code{Spectral}.}

\item{legend_title}{The title for the legend.}

\item{...}{Other arguments currently ignored.}
}
\description{
This function creates a Hovmoller plot for either latitude or longitude.
}
\examples{
# dataframe example
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
  proc == "Tmax" &
  month \%in\% 5:9 &
  year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
hov <- hovmoller(lat_or_lon = "lat",
          x = Tmax,
          lat_or_lon_col = 'lat',
          t_col = 't',
          z_col = 'z')
autoplot(hov, legend_title = 'Temperature')


# stars example
library(stars)
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
hov <- hovmoller(prec)
autoplot(hov, ylab = "Time", xlab = "Latitude")
}
