% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_semivariogram.R
\name{autoplot.semivariogramobj}
\alias{autoplot.semivariogramobj}
\title{Plots the semi-variogram computed from the semivariogram algorithm}
\usage{
\method{autoplot}{semivariogramobj}(object, ...)
}
\arguments{
\item{object}{The output from the semivariogram function.}

\item{...}{Other arguments currently ignored.}
}
\description{
A semivariogram plot
}
\examples{
library(dplyr)
data(locs)
data(Times)
data(Tmax)
temp_part <- with(Times, paste(year, month, day, sep = "-"))
temp_part <- data.frame(date = as.Date(temp_part)[913:943])
Tmax <- Tmax[913:943, ]
semiv <- semivariogram(locs,
        temp_part,
        Tmax,
        latitude_linear = FALSE,
        longitude_linear = FALSE,
        missing_value = -9999,
        width = 50,
        cutoff = 1000,
        tlagmax = 7 )
autoplot(semiv)

}
