% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_emp_spatial_cov.R
\name{autoplot.spatialcov}
\alias{autoplot.spatialcov}
\title{Plots empirical spatial covariance using emp_spatial_cov function}
\usage{
\method{autoplot}{spatialcov}(object, xlab = "Latitude", ...)
}
\arguments{
\item{object}{The output of the function `emp_spatial_cov'.}

\item{xlab}{The label for x-axis.}

\item{...}{Other arguments currently ignored.}
}
\description{
You can specify up to 4 latitude or longitude strips in your plot.
This lets you split the spatial domain for better visualization.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
               proc == "Tmax" &
               month \%in\% 5:9 &
               year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
esv <- emp_spatial_cov(Tmax,
                       lat_col = "lat",
                       lon_col = "lon",
                       t_col ="t",
                       z_col = "z",
                       lat_or_lon_strips = "lon",
                       num_strips = 4,
                       lag = 1)
autoplot(esv)
}
