% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_temporal_means.R
\name{autoplot.temporalmeans}
\alias{autoplot.temporalmeans}
\title{Plots temporal empirical means computed from temporal_means function}
\usage{
\method{autoplot}{temporalmeans}(
  object,
  ylab = "Mean Value",
  xlab = "Month",
  legend_title = "",
  title = "Temporal Empirical Means",
  ...
)
}
\arguments{
\item{object}{The output of the function `temporal_means'.}

\item{ylab}{The y label.}

\item{xlab}{The x label.}

\item{legend_title}{The title for the legend.}

\item{title}{The graph title.}

\item{...}{Other arguments currently ignored.}
}
\description{
A line plot of temporal means with observed data in the background.
}
\examples{
# dataframe example
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
temp_means <- temporal_means(Tmax,
              t_col = 'date',
              z_col = 'z',
              id_col = 'id')
autoplot(temp_means)

# stars example
library(stars)
library(dplyr)
library(units)
# Example
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
temp_means <- temporal_means(prec)
# autoplot(temp_means)
}
