% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_correlation.R
\name{canonical_correlation.data.frame}
\alias{canonical_correlation.data.frame}
\title{Computes transformed variables from Canonical Correlation Analysis using a dataframe}
\usage{
\method{canonical_correlation}{data.frame}(x1, x2, ...)
}
\arguments{
\item{x1}{The first dataframe or stars object.}

\item{x2}{The second dataframe or stars objext. The dimensions of both datasets need to be the same.}

\item{...}{Other arguments that need to be used for datafames or currently ignored.}
}
\description{
Computes Canonical Correlation Analysis (CCA) using 2 datasets.
}
\examples{
df1 <- SSTdatashort[1:100, ]
df2 <- SSTdatashort[401:500, ]
canonical_correlation(df1, df2)

}
