% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_orth_fun.R
\name{emp_orth_fun.stars}
\alias{emp_orth_fun.stars}
\title{Computes empirical orthogonal functions for a stars object}
\usage{
\method{emp_orth_fun}{stars}(x, ...)
}
\arguments{
\item{x}{The dataframe or stars object. If it is a dataframe, then it should have the locations.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
Computes empirical orthogonal functions of the data.
}
\examples{
library(dplyr)
library(stars)
# Create a stars object from a data frame
precip_df <- NOAA_df_1990[NOAA_df_1990$proc == 'Precip', ] \%>\%
  filter(date >= "1992-02-01" & date <= "1992-02-28")
precip <- precip_df[ ,c('lat', 'lon', 'date', 'z')]
st_precip <- st_as_stars(precip, dims = c("lon", "lat", "date"))
emp_orth_fun(st_precip)

}
