% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_spatial_cov.R
\name{emp_spatial_cov.stars}
\alias{emp_spatial_cov.stars}
\title{Computes empirical spatial covariance using a stars object}
\usage{
\method{emp_spatial_cov}{stars}(
  x,
  lat_or_lon_strips = "lon",
  quadratic_time = FALSE,
  quadratic_space = FALSE,
  num_strips = 1,
  lag = 0,
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{lat_or_lon_strips}{Takes the values \code{lat} or \code{lon}. The value \code{lat} produces latitudinal strips,
i.e., covariance plots over longitude for different latitudinal strips. The value \code{lon} produces longitudinal
strips, i.e., covariance plots over latitude for different longitudinal strips.}

\item{quadratic_time}{If \code{TRUE}  a linear model with quadratic time is fitted and residuals computed. If \code{FALSE}
the model is fitted with linear space and time coefficients.}

\item{quadratic_space}{If \code{TRUE}  a linear model with quadratic space is fitted and residuals computed. If \code{FALSE}
the model is fitted with linear space and time coefficients.}

\item{num_strips}{The number of latitudinal/longitudinal strips to produce. This is used when plotting using autoplot.}

\item{lag}{Lag can be either 0 or 1.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
Computes empirical spatial covariance by removing trends and examining residuals. It can compute lag-0 or log-1
empirical covariance either by latitude or longitude. You can split up the spatial domain by latitude or
longitude and plot the covariance for each longitudinal/latitudinal strips.
}
\examples{
\dontrun{
library(stars)
library(dplyr)
library(units)
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
prec2 <- prec \%>\%
         slice(time,  1:3)
emp_spatial_cov(prec2,
                lat_or_lon_strips = "lon",
                num_strips = 4,
                lag = 1)
}
}
