% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_hovmoller.R
\name{hovmoller}
\alias{hovmoller}
\title{Generic function to compute the data structure for Hovmoller plots}
\usage{
hovmoller(x, lat_or_lon = "lat", xlen = NULL, ...)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{lat_or_lon}{Needs to be either \code{lat} or \code{lon}. \code{lat} plots the latitudinal
Hovmoller plat, while \code{lon} plots the longitudinal Hovmoller plot.}

\item{xlen}{The length of the xaxis for latitude/longitude.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
This function creates the data structure for Hovmoller plots for either latitude or longitude.
This function can take either a stars object or a dataframe. Input arguments differ for each case.
}
\examples{
# dataframe examples
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
  proc == "Tmax" &
  month \%in\% 5:9 &
  year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
hovmoller(lat_or_lon = "lat",
          x = Tmax,
          lat_or_lon_col = 'lat',
          t_col = 't',
          z_col = 'z')

# stars examples
library(stars)
}
