% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs.R
\docType{data}
\name{locs}
\alias{locs}
\title{The locations used in the NOAA dataset.}
\format{
A data frame with 328 rows and 3 variables:
\describe{
  \item{id}{Location is}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
  ...
}
}
\usage{
locs
}
\description{
This dataset is included in the STRbook R package.
}
\keyword{datasets}
