% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline.R
\name{ridgeline.data.frame}
\alias{ridgeline.data.frame}
\title{Ridgeline plots grouped by an attribute using a dataframe as an input.}
\usage{
\method{ridgeline}{data.frame}(
  x,
  num_grps = 10,
  xlab = "Value",
  ylab = "Group Intervals",
  title = "",
  legend_title = "z",
  group_col,
  z_col,
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{num_grps}{The number of levels for the ridgeline plot.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{legend_title}{The title for the legend.}

\item{group_col}{The column name of the group column.}

\item{z_col}{For dataframes: the The quantity of interest that will be plotted. Eg. temperature.}

\item{...}{Other arguments currently ignored.}
}
\description{
Plots ridgeline plots grouped by latitude/longitude or time.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
TmaxJan <- filter(NOAA_df_1990,
                 proc == "Tmax" &
                 year == 1993 &
                 month == 1)
ridgeline(TmaxJan,
      group_col = 'lat',
      z_col = 'z',
      xlab = 'Maximum Temperature',
      ylab = 'Latitude Intervals')

}
