% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline.R
\name{ridgeline.stars}
\alias{ridgeline.stars}
\title{Ridgeline plots grouped by an attribute using a stars object as input.}
\usage{
\method{ridgeline}{stars}(
  x,
  num_grps = 10,
  xlab = "Value",
  ylab = "Group Intervals",
  title = "",
  legend_title = "z",
  group_dim,
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{num_grps}{The number of levels for the ridgeline plot.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{legend_title}{The title for the legend.}

\item{group_dim}{The dimension for the grouping variable.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
Plots ridgeline plots grouped by latitude/longitude or time.
}
\examples{
library(stars)
library(dplyr)
library(units)

# Example 1
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
ridgeline(prec, group_dim = 1)


# Example 2
tif = system.file("tif/olinda_dem_utm25s.tif", package = "stars")
x <- read_stars(tif)
dim(x)
ridgeline(x, group_dim = 1)
ridgeline(x, group_dim = 2)


# Example 3
tif = system.file("tif/lc.tif", package = "stars")
x <- read_stars(tif)
ridgeline(x, group_dim = 1)
ridgeline(x, group_dim = 2)

}
