% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semivariogram.R
\name{semivariogram.stars}
\alias{semivariogram.stars}
\title{Computes the semi-variogram using a stars object}
\usage{
\method{semivariogram}{stars}(
  x,
  latitude_linear = TRUE,
  longitude_linear = TRUE,
  missing_value = -9999,
  width = 80,
  cutoff = 1000,
  tlagmax = 6,
  ...
)
}
\arguments{
\item{x}{The dataframe or stars object. If it is a dataframe, then it should have the locations.}

\item{latitude_linear}{If \code{TRUE}  a linear model is fitted with latitude as a covariate is fitted.}

\item{longitude_linear}{If \code{TRUE}  a linear model is fitted with longitude as a covariate is fitted.}

\item{missing_value}{If a certain value such as -9999 denotes the missing values for given locations and times.}

\item{width}{A parameter to the \code{gstat::variogram} function. The width of the distance intervals to be considered.}

\item{cutoff}{A parameter to the \code{gstat::variogram} function. The spatial separation distance.}

\item{tlagmax}{A parameter to the \code{gstat::variogram} function. The maximum time lag.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
Computes the semi-variogram using a stars object. A 3 dimensional stars object is used with
the first 2 for locations and the third for time.
}
\examples{
library(dplyr)
data(locs)
data(Times)
data(Tmax)
temp_part <- with(Times, paste(year, month, day, sep = "-"))
temp_part <- data.frame(date = as.Date(temp_part)[913:943])
Tmax <- Tmax[913:943, ]
semivariogram(locs,
              latitude_linear = FALSE,
              longitude_linear = FALSE,
              missing_value = -9999,
              width = 50,
              cutoff = 1000,
              tlagmax = 7,
              times_df = temp_part,
              values_df = Tmax
)

}
