% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_means.R
\name{spatial_means}
\alias{spatial_means}
\title{Computes spatial empirical means}
\usage{
spatial_means(x, lat_col, lon_col, t_col, z_col)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{lat_col}{The column or the column name giving the latitude. The y coordinate can be used instead of latitude.}

\item{lon_col}{The column or the column name giving the longitude. The x coordinate can be used instead of longitude.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}
}
\description{
This function computes spatial empirical means by latitude and longitude averaged over time.
}
\examples{
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
spatial_means(Tmax,
       lat_col = "lat",
       lon_col = "lon",
       t_col = "t",
       z_col = "z")

}
