% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_spatial_snapshots.R
\name{spatial_snapshots}
\alias{spatial_snapshots}
\title{Generic function to plot spatial snapshots of data through time}
\usage{
spatial_snapshots(
  x,
  xlab = "x",
  ylab = "y",
  title = "",
  palette = "Spectral",
  legend_title = "z",
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{palette}{The color palette. Default is \code{Spectral}.}

\item{legend_title}{The title for the legend.}

\item{...}{Other arguments that need to be used for datafames to specify columns.}
}
\value{
A ggplot object.
}
\description{
This function can take either a stars object or a dataframe. Input arguments differ
for each case.
}
\examples{
library(dplyr)
# Dataframe example
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
  proc == "Tmax" &
  month \%in\% 5:9 &
  year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
Tmax_days <- subset(Tmax, t \%in\% c(1, 15, 30))
spatial_snapshots(Tmax_days,
  lat_col = 'lat',
  lon_col = 'lon',
  t_col = 't',
  z_col = 'z',
  title = "Maximum Temperature for 3 days ")

# stars example
library(stars)
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x <- read_stars(tif)
spatial_snapshots(x)

}
\seealso{
spatial_snapshots.data.frame

spatial_snapshots.stars
}
