% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_snapshots.R
\name{spatial_snapshots.stars}
\alias{spatial_snapshots.stars}
\title{Plots spatial snapshots of data through time using a stars object as input.}
\usage{
\method{spatial_snapshots}{stars}(
  x,
  xlab = "x",
  ylab = "y",
  title = "",
  palette = "Spectral",
  legend_title = "z",
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{palette}{The color palette. Default is \code{Spectral}.}

\item{legend_title}{The title for the legend.}

\item{...}{Other arguments that need to be used for datafames to specify columns.}
}
\value{
A ggplot object.
}
\description{
This function plots a set of spatial snapshots through time.
}
\examples{
library(stars)
library(dplyr)
# The third dimension is not time, but bands in this example
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x <- read_stars(tif)
spatial_snapshots(x)

# The third dimension is time in this example
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
prec2 <- prec \%>\% slice(index =c(1,3,5), along = time)
spatial_snapshots(prec2)

}
