% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTdata.R
\docType{data}
\name{SSTdatashort}
\alias{SSTdatashort}
\title{The data from of the Sea Surface Temperature (SST) dataset. A subset of the original dataset is used.}
\format{
A dataframe with 500 rows and 396 columns.
}
\usage{
SSTdatashort
}
\description{
The original dataset is included in the STRbook R package.
}
\keyword{datasets}
