% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTlandmask.R
\docType{data}
\name{SSTlandmaskshort}
\alias{SSTlandmaskshort}
\title{The land mask for the Sea Surface Temperature (SST) dataset. A subset of the original dataset is used.}
\format{
A dataframe with 500 rows and 1 column.
\describe{
  \item{mask}{A value of 1 is given if the location covers land.}
  ...
}
}
\usage{
SSTlandmaskshort
}
\description{
The original dataset is included in the STRbook R package.
}
\keyword{datasets}
