% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTlonlat.R
\docType{data}
\name{SSTlonlatshort}
\alias{SSTlonlatshort}
\title{The locations of the Sea Surface Temperatures (SST) dataset. A subset of the original dataset is used.}
\format{
A data frame with 500 rows and 2 variables:
\describe{
  \item{lon}{Longitude}
  \item{lat}{Latitude}
  ...
}
}
\usage{
SSTlonlatshort
}
\description{
The original dataset is included in the STRbook R package.
}
\keyword{datasets}
