% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancor.R
\name{st_cancor}
\alias{st_cancor}
\title{Plots transformed variables from Canonical Correlation Analysis}
\usage{
st_cancor(df1, df2, xlab = "Year")
}
\arguments{
\item{df1}{The first dataset.}

\item{df2}{The second dataset. The dimensions of both datasets need to be the same.}

\item{xlab}{The xlabel to appear on CCA plot.}
}
\description{
Performs Canonical Correlation Analysis (CCA) using 2 datasets and plots the first CC function of both datasets.
}
\examples{
df1 <- SSTdatashort[1:100, ]
df2 <- SSTdatashort[401:500, ]
st_cancor(df1, df2, xlab = "Longitude")

}
