% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancor_eof.R
\name{st_cancor_eof}
\alias{st_cancor_eof}
\title{Plots Canonical Correlation Analysis plots using Empirical Orthogonal Functions from a lagged dataset}
\usage{
st_cancor_eof(
  locations_df,
  values_df,
  lag = 7,
  n_eof = 10,
  xlab = "Year",
  line_plot = TRUE,
  space_plot = TRUE,
  palette = "Spectral"
)
}
\arguments{
\item{locations_df}{The dataframe containing latitude and longitude (colnames lat and lon).}

\item{values_df}{The dataframe of dimension \code{length(times) x length(locations)} containing the quantity of interest.}

\item{lag}{Specifies the lag to be used.}

\item{n_eof}{The number of EOFs to be used.}

\item{xlab}{The x label for the line plot.}

\item{line_plot}{If set to \code{TRUE}, then the line plot is included.}

\item{space_plot}{If set to \code{TRUE}, the space splot is included.}

\item{palette}{The color palette. Default is \code{Spectral}.}
}
\description{
Plots CCA plots using EOF analysis using a lagged dataset.
}
\examples{
st_cancor_eof(SSTlonlatshort,
              SSTdatashort,
              n_eof = 8)

}
