% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecov.R
\name{st_ecov}
\alias{st_ecov}
\title{Plots empirical spatial covariance}
\usage{
st_ecov(
  df,
  lat_col,
  lon_col,
  t_col,
  z_col,
  lat_or_lon_strips = "lon",
  quadratic_time = TRUE,
  quadratic_space = TRUE,
  lag = 0,
  num_strips = 1,
  xlab = "Latitude"
)
}
\arguments{
\item{df}{The input data in a dataframe.}

\item{lat_col}{The column or the column name giving the latitude. The y coordinate can be used instead of latitude.}

\item{lon_col}{The column or the column name giving the longitude. The x coordinate can be used instead of longitude.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{lat_or_lon_strips}{Takes the values \code{lat} or \code{lon}. The value \code{lat} produces latitudinal strips,
i.e., covariance plots over longitude for different latitudinal strips. The value \code{lon} produces longitudinal
strips, i.e., covariance plots over latitude for different longitudinal strips.}

\item{quadratic_time}{If \code{TRUE}  a linear model with quadratic time is fitted and residuals computed. If \code{FALSE}
the model is fitted with linear space and time coefficients.}

\item{quadratic_space}{If \code{TRUE}  a linear model with quadratic space is fitted and residuals computed. If \code{FALSE}
the model is fitted with linear space and time coefficients.}

\item{lag}{Lag can be either 0 or 1.}

\item{num_strips}{The number of latitudinal/longitudinal strips to produce. This is the number of graphs you get.}

\item{xlab}{The x label.}
}
\description{
Plots empirical spatial covariance by removing trends and examining residuals. It can plot lag-0 or log-1
empirical covariance either by latitude or longitude. You can split up the spatial domain by latitude or
longitude and plot the covariance for each longitudinal/latitudinal strips.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
  proc == "Tmax" &
  month \%in\% 5:9 &
  year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
st_ecov(Tmax,
        lat_col = "lat",
        lon_col = "lon",
        t_col ="t",
        z_col = "z",
        num_strips = 4,
        lat_or_lon_strips = "lon",
        lag = 1)

}
