% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl.R
\name{st_rl}
\alias{st_rl}
\title{Ridgeline plots grouped by an attribute.}
\usage{
st_rl(
  df,
  group_col,
  z_col,
  num_grps = 10,
  xlab = "Value",
  ylab = "Group Intervals",
  title = "",
  legend_title = "z"
)
}
\arguments{
\item{df}{The input data in a dataframe.}

\item{group_col}{The column name of the group column.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{num_grps}{The number of levels for the ridgeline plot.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{legend_title}{The title for the legend.}
}
\description{
Plots ridgeline plots grouped by latitude/longitude or time.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
TmaxJan <- filter(NOAA_df_1990,
                 proc == "Tmax" &
                 year == 1993 &
                 month == 1)
st_rl(TmaxJan,
      group_col = 'lat',
      z_col = 'z',
      xlab = 'Maximum Temperature',
      ylab = 'Latitude Intervals')

}
