% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{st_sem}
\alias{st_sem}
\title{Plots spatial empirical means}
\usage{
st_sem(df, lat_col, lon_col, t_col, z_col, ylab = "Mean Value")
}
\arguments{
\item{df}{The input data in a dataframe.}

\item{lat_col}{The column or the column name giving the latitude. The y coordinate can be used instead of latitude.}

\item{lon_col}{The column or the column name giving the longitude. The x coordinate can be used instead of longitude.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{ylab}{The y label.}
}
\description{
This function plots two graphs, spatial empirical means by latitude and longitude.
}
\examples{
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
st_sem(Tmax,
       lat_col = "lat",
       lon_col = "lon",
       t_col = "t",
       z_col = "z",
       ylab="Mean Max Temp")

}
