% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiv.R
\name{st_semiv}
\alias{st_semiv}
\title{Plots the semi-variogram}
\usage{
st_semiv(
  locations_df,
  times_df,
  values_df,
  latitude_linear = TRUE,
  longitude_linear = TRUE,
  missing_value = -9999,
  width = 80,
  cutoff = 1000,
  tlagmax = 6
)
}
\arguments{
\item{locations_df}{The dataframe containing latitude and longitude (colnames lat and lon).}

\item{times_df}{The dataframe containing the dates in \code{Date} format.}

\item{values_df}{The dataframe of dimension \code{length(times) x length(locations)} containing the quantity of interest.}

\item{latitude_linear}{If \code{TRUE}  a linear model is fitted with latitude as a covariate is fitted.}

\item{longitude_linear}{If \code{TRUE}  a linear model is fitted with longitude as a covariate is fitted.}

\item{missing_value}{If a certain value such as -9999 denotes the missing values for given locations and times.}

\item{width}{A parameter to the \code{gstat::variogram} function. The width of the distance intervals to be considered.}

\item{cutoff}{A parameter to the \code{gstat::variogram} function. The spatial separation distance.}

\item{tlagmax}{A parameter to the \code{gstat::variogram} function. The maximum time lag.}
}
\description{
Plots the semi-variogram given the locations, times and the quantity of interest.
}
\examples{
library(dplyr)
data(locs)
data(Times)
data(Tmax)
temp_part <- with(Times, paste(year, month, day, sep = "-"))
temp_part <- data.frame(date = as.Date(temp_part)[913:943])
Tmax <- Tmax[913:943, ]
st_semiv(locs,
        temp_part,
        Tmax,
        latitude_linear = FALSE,
        longitude_linear = FALSE,
        missing_value = -9999,
        width = 50,
        cutoff = 1000,
        tlagmax = 7
)

}
