% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{st_tem}
\alias{st_tem}
\title{Plots temporal empirical means}
\usage{
st_tem(
  df,
  t_col,
  z_col,
  id_col,
  ylab = "Mean Value",
  xlab = "Month",
  legend_title = "",
  title = "Temporal Empirical Means"
)
}
\arguments{
\item{df}{The input data in a dataframe.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{id_col}{The column of the location id.}

\item{ylab}{The y label.}

\item{xlab}{The x label.}

\item{legend_title}{The title for the legend.}

\item{title}{The graph title.}
}
\description{
This function plots temporal empirical means averaged per time unit.
}
\examples{
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
st_tem(Tmax,
       t_col = 'date',
       z_col = 'z',
       id_col = 'id')
}
