% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stxplore-package.R
\docType{package}
\name{stxplore-package}
\alias{stxplore-package}
\title{stxplore: Exploration of Spatio-Temporal Data}
\description{
A set of statistical tools for spatio-temporal data exploration. Includes simple plotting functions, covariance calculations and computations similar to principal component analysis for spatio-temporal data. Can use both dataframes and stars objects for all plots and computations. For more details refer 'Spatio-Temporal Statistics with R' (Christopher K. Wikle, Andrew Zammit-Mangion, Noel Cressie, 2019, ISBN:9781138711136).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/stxplore/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Petra Kunhert \email{petra.kuhnert@data61.csiro.au} (\href{https://orcid.org/0000-0001-9070-0091}{ORCID})
}

Other contributors:
\itemize{
  \item Andrew Zammit-Mangion \email{azm@uow.edu.au} (\href{https://orcid.org/0000-0002-4164-6866}{ORCID}) [contributor]
}

}
\keyword{internal}
