% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_means.R
\name{temporal_means.data.frame}
\alias{temporal_means.data.frame}
\title{Computes temporal empirical means using a dataframe as input}
\usage{
\method{temporal_means}{data.frame}(x, t_col, z_col, id_col, ...)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{t_col}{For dataframes: the time column. Time must be a set of discrete integer values.}

\item{z_col}{For dataframes: the The quantity of interest that will be plotted. Eg. temperature.}

\item{id_col}{The column of the location id.}

\item{...}{Other arguments currently ignored.}
}
\description{
This function computes temporal empirical means averaged per time unit.
}
\examples{
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
temporal_means(Tmax,
       t_col = 'date',
       z_col = 'z',
       id_col = 'id')
}
