% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_means.R
\name{temporal_means.stars}
\alias{temporal_means.stars}
\title{Computes temporal empirical means using a stars object as input}
\usage{
\method{temporal_means}{stars}(x, ...)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{...}{Other arguments currently ignored.}
}
\description{
This function computes temporal empirical means averaged per time unit.
}
\examples{
library(stars)
library(dplyr)
library(units)
# Example
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
temporal_means(prec)
}
