% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_snapshot.R
\name{temporal_snapshots.data.frame}
\alias{temporal_snapshots.data.frame}
\title{Plots temporal snapshots of data for specific spatial locations using a dataframe as input.}
\usage{
\method{temporal_snapshots}{data.frame}(
  x,
  xlab = "Time",
  ylab = "Value",
  title = "",
  t_col,
  z_col,
  id_col,
  id_sample,
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{id_col}{The column of the location id.}

\item{id_sample}{The sample of location ids to be plotted}

\item{...}{Other arguments currently ignored.}
}
\description{
This function plots temporal snapshos for specific spatial locations. The location id sample
need to be given as a function argument.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
             proc == "Tmax" &
             month \%in\% 5:9 &
             year == 1993)
Tmax_ID <- unique(Tmax$id)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
ids <- sample(Tmax_ID, 10)
temporal_snapshots(Tmax,
                  t_col = 't',
                  z_col = 'z',
                  id_col = 'id',
                  id_sample = ids)
}
