% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_snapshot.R
\name{temporal_snapshots.stars}
\alias{temporal_snapshots.stars}
\title{Plots temporal snapshots of data for specific spatial locations using a stars object as input.}
\usage{
\method{temporal_snapshots}{stars}(
  x,
  xlab = "Time",
  ylab = "Value",
  title = "",
  xvals,
  yvals,
  precision = 0,
  ...
)
}
\arguments{
\item{x}{A stars object or a dataframe. Arguments differ according to the input type.}

\item{xlab}{The x label.}

\item{ylab}{The y label.}

\item{title}{The graph title.}

\item{xvals}{The set of xvalues to plot.}

\item{yvals}{The set of yvalues to plot. These two lengths need to be the same.}

\item{precision}{Set to 0, if the given values are compared with the integer values in the stars object.}

\item{...}{Other arguments currently ignored.}
}
\description{
This function plots temporal snapshos for specific spatial locations. The location x and y
indices need to be given.
}
\examples{
library(stars)
tif = system.file("tif/L7_ETMs.tif", package = "stars")
x <- read_stars(tif)
xvals <- c(288876.0,289047.0)
yvals <- c(9120405, 9120006)
temporal_snapshots(x,
                  xvals = xvals,
                  yvals = yvals)
}
