% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\docType{data}
\name{Times}
\alias{Times}
\title{The time period in which the NOAA dataset was recorded. This spans from January 1990 to  December 1993.}
\format{
A data frame with 1461 rows and 4 variables:
\describe{
  \item{julian}{Day in Julian time}
  \item{year}{The year}
  \item{month}{The month}
  \item{day}{The day}
  ...
}
}
\usage{
Times
}
\description{
This dataset is included in the STRbook R package.
}
\keyword{datasets}
