% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmax.R
\docType{data}
\name{Tmax}
\alias{Tmax}
\title{The maximum temperature values used in the NOAA dataset in a wide dataframe format.}
\format{
A data frame with 1461 rows and columns having maximum temperature
for times and locations in data locs and Times.
}
\usage{
Tmax
}
\description{
This dataset is included in the STRbook R package.
}
\keyword{datasets}
