% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_spatial_means.R
\name{autoplot.spatialmeans}
\alias{autoplot.spatialmeans}
\title{Plots spatial empirical means computed from spatial_means function}
\usage{
\method{autoplot}{spatialmeans}(object, ylab = "Mean Value", ...)
}
\arguments{
\item{object}{The output from the `spatial_means' function.}

\item{ylab}{The label for y axis.}

\item{...}{Other arguments currently ignored.}
}
\description{
This function computes spatial empirical means by latitude and longitude averaged over time.
}
\examples{
data(NOAA_df_1990)
library(dplyr)
Tmax <- filter(NOAA_df_1990,                      # subset the data
              proc == "Tmax" &                   # extract max temperature
                month \%in\% 5:9 &                 # May to July
                year == 1993)                    # year 1993
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1      # create a new time variable starting at 1
sem <- spatial_means(Tmax,
       lat_col = "lat",
       lon_col = "lon",
       t_col = "t",
       z_col = "z")
autoplot(sem, ylab="Mean Max Temp")
}
