% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancor_eof.R
\name{cancor_eof}
\alias{cancor_eof}
\title{Performs Canonical Correlation Analysis  using Empirical Orthogonal Functions from a lagged dataset}
\usage{
cancor_eof(x, values_df, lag = 7, n_eof = 10)
}
\arguments{
\item{x}{The dataframe or stars object. If it is a dataframe, then it should have the locations.}

\item{values_df}{The dataframe of dimension \code{length(times) x length(locations)} containing the quantity of interest.}

\item{lag}{Specifies the lag to be used.}

\item{n_eof}{The number of EOFs to be used.}
}
\description{
Performs CCA using EOF analysis using a lagged dataset.
}
\examples{
cancor_eof(SSTlonlatshort,
           SSTdatashort,
           lag = 7,
           n_eof = 8)

}
