% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_cancor_eof.R
\name{cancor_eof}
\alias{cancor_eof}
\title{Generic function to perform Canonical Correlation Analysis  using Empirical Orthogonal Functions from a lagged dataset}
\usage{
cancor_eof(x, lag, n_eof, ...)
}
\arguments{
\item{x}{The dataframe or stars object. If it is a dataframe, then it should have the locations.}

\item{lag}{Specifies the lag to be used.}

\item{n_eof}{The number of EOFs to be used.}

\item{...}{Other arguments that need to be used for datafames  or currently ignored.}
}
\description{
Performs CCA using EOF analysis using a lagged dataset.
}
\examples{
# Dataframe example
cancor_eof(x = SSTlonlatshort,
           lag = 7,
           n_eof = 8,
           values_df = SSTdatashort)

# Stars example
library(dplyr)
library(stars)
# Create a stars object from a data frame
precip_df <- NOAA_df_1990[NOAA_df_1990$proc == 'Precip', ] \%>\%
  filter(date >= "1992-02-01" & date <= "1992-02-28")
precip <- precip_df[ ,c('lat', 'lon', 'date', 'z')]
st_precip <- st_as_stars(precip, dims = c("lon", "lat", "date"))
cancor_eof(st_precip)
}
