% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_correlation.R
\name{canonical_correlation}
\alias{canonical_correlation}
\title{Computes transformed variables from Canonical Correlation Analysis}
\usage{
canonical_correlation(df1, df2)
}
\arguments{
\item{df1}{The first dataset.}

\item{df2}{The second dataset. The dimensions of both datasets need to be the same.}
}
\description{
Computes Canonical Correlation Analysis (CCA) using 2 datasets.
}
\examples{
df1 <- SSTdatashort[1:100, ]
df2 <- SSTdatashort[401:500, ]
canonical_correlation(df1, df2)

}
