% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_orth_fun.R
\name{emp_orth_fun}
\alias{emp_orth_fun}
\title{Computes empirical orthogonal functions}
\usage{
emp_orth_fun(locations_df, values_df)
}
\arguments{
\item{locations_df}{The dataframe containing latitude and longitude (colnames lat and lon).}

\item{values_df}{The dataframe of dimension \code{length(times) x length(locations)} containing the quantity of interest.}
}
\description{
Computes empirical orthogonal functions of the data.
}
\examples{
data(SSTlonlatshort)
data(SSTdatashort)
data(SSTlandmaskshort)
# Take first 396 months (33 years) and delete land
delete_rows <- which(SSTlandmaskshort  ==  1)
SSTdatashort   <- SSTdatashort[-delete_rows, 1:396]
emp_orth_fun(SSTlonlatshort[-delete_rows,  ],
             SSTdatashort)

}
