% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hovmoller.R
\name{hovmoller}
\alias{hovmoller}
\title{Computes the data structure for the Hovmoller plots}
\usage{
hovmoller(lat_or_lon = "lat", df, lat_or_lon_col, t_col, z_col, xlen = NULL)
}
\arguments{
\item{lat_or_lon}{Needs to be either \code{lat} or \code{lon}. \code{lat} plots the latitudinal
Hovmoller plat, while \code{lon} plots the longitudinal Hovmoller plot.}

\item{df}{The input data in a dataframe.}

\item{lat_or_lon_col}{The column or the column name corresponding to the latitude/longitude.}

\item{t_col}{The time column. Time must be a set of discrete integer values.}

\item{z_col}{The The quantity of interest that will be plotted. Eg. temperature.}

\item{xlen}{The length of the xaxis for latitude/longitude.}
}
\description{
This function creates the data structure for Hovmoller plots for either latitude or longitude.
}
\examples{
library(dplyr)
data(NOAA_df_1990)
Tmax <- filter(NOAA_df_1990,
  proc == "Tmax" &
  month \%in\% 5:9 &
  year == 1993)
Tmax$t <- Tmax$julian - min(Tmax$julian) + 1
hovmoller(lat_or_lon = "lat",
          df = Tmax,
          lat_or_lon_col = 'lat',
          t_col = 't',
          z_col = 'z')

}
