% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stxplore-package.R
\docType{package}
\name{stxplore-package}
\alias{stxplore-package}
\title{stxplore: Exploration of Spatio-Temporal Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/stxplore/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

Authors:
\itemize{
  \item Petra Kunhert \email{Petra.Kuhnert@data61.csiro.au>} (\href{https://orcid.org/0000-0001-9070-0091}{ORCID})
}

Other contributors:
\itemize{
  \item Andrew Zammit-Mangion \email{azm@uow.edu.au>} (\href{https://orcid.org/0000-0002-4164-6866}{ORCID}) [contributor]
}

}
\keyword{internal}
